make index type int

+++ rqpy/process/_process_rq.py
@@ -1575,7 +1575,7 @@ def _calc_rq_single_channel(signal, template, psd, setup, readout_inds, chan, ch
 
             if setup.do_ofamp_constrained[chan_num]:
                 if setup.ofamp_constrained_usetrigsimcenter:
-                    windowcenter_constrain = triggertime_sim[jj] * setup.fs - (signal.shape[-1]//2)
+                    windowcenter_constrain = int(triggertime_sim[jj] * setup.fs - (signal.shape[-1]//2))
                     if setup.indstart is not None:
                         windowcenter_constrain -= setup.indstart
                 else:
